refactored code into an extra file x2
