+++ src/executor/executor.c
@@ -32,8 +32,7 @@ int	exec_selector(t_data *core, t_command *command)
 	}
 	else
 	{
-		retcode = execve(cmdpath, args, envp);
-		if (retcode)
+		if (execve(cmdpath, args, envp))
 			perror(args[0]);
 		free(cmdpath);
 	}
